fastseo/requirements.txt
fastapi
uvicorn
celery
typing
redis
redisbeat
amqp
billiard
clickhouse-driver
python-dotenv==0.12.0